transpose WIP

+++ nstream-kernel-implementations.F90
@@ -1,6 +1,6 @@
 subroutine nstream_length(length,scalar,A,B,C) bind(C)
     use, intrinsic :: iso_fortran_env
-    integer(kind=REAL64), value :: length
+    integer(kind=INT64), value :: length
     real(kind=REAL64), value :: scalar
     real(kind=REAL64), dimension(length) :: A,B,C
     do concurrent (i=1:length)
@@ -10,7 +10,7 @@ end subroutine nstream_length
 
 subroutine nstream_star(length,scalar,A,B,C) bind(C)
     use, intrinsic :: iso_fortran_env
-    integer(kind=REAL64), value :: length
+    integer(kind=INT64), value :: length
     real(kind=REAL64), value :: scalar
     real(kind=REAL64), dimension(*) :: A,B,C
     do concurrent (i=1:length)
@@ -20,7 +20,7 @@ end subroutine nstream_star
 
 subroutine nstream_colon(length,scalar,A,B,C) bind(C)
     use, intrinsic :: iso_fortran_env
-    integer(kind=REAL64), value :: length
+    integer(kind=INT64), value :: length
     real(kind=REAL64), value :: scalar
     real(kind=REAL64), dimension(:) :: A,B,C
 #if 0
